6 size

+++ 0x00-hello_world/6-size.c
@@ -13,10 +13,10 @@ int main(void)
 	long long int d;
 	float f;
 
-printf("Size of a char: %lu bytes(s)\n", sizeof(char));
-printf("Size of an int: %lu bytes(s)\n", sizeof(int));
-printf("Size of a long int: %lu bytes(s)\n", sizeof(long int));
-printf("Size of a long long int: %lu bytes(s)\n", sizeof(long long int));
-printf("Size of float: %lu bytes(s)\n", sizeof(float));
+printf("Size of a char: %lu bytes(s)\n", (unsigned long)sizeof(char));
+printf("Size of an int: %lu bytes(s)\n",(unsigned long)sizeof(int));
+printf("Size of a long int: %lu bytes(s)\n",(unsigned long)sizeof(long int));
+printf("Size of a long long int: %lu bytes(s)\n",(unsigned long)sizeof(long long int));
+printf("Size of float: %lu bytes(s)\n",(unsigned long)sizeof(float));
 return (0);
 }
